+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -2,10 +2,19 @@
 
 ## Unreleased
 
+## 1.8.0-rc.2
+
+Released 2024-May-13
+
 * **Experimental (pre-release builds only)**: Add support for exporting
   exemplars when OTLP protobuf encoding is enabled via
   `PrivatePreviewEnableOtlpProtobufEncoding=true` in the connection string.
   ([#1703](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1703))
+
+* Add support for exporting
+  exponential histograms when OTLP protobuf encoding is enabled via
+  `PrivatePreviewEnableOtlpProtobufEncoding=true` in the connection string.
+  ([#1705](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1705))
 
 ## 1.8.0-rc.1